integration tests: authentication, getAll, getMeter, delete

+++ src/main/java/com/dima/meterscollector/controller/MeterConfigController.java
@@ -76,6 +76,9 @@ public class MeterConfigController {
 
     @DeleteMapping("/{id}")
     public void deleteMeter(@PathVariable long id){
+        if(!meterConfigRepo.existsById(id)){
+            throw new NotFoundException();
+        }
         meterConfigRepo.deleteById(id);
         pollMeters.setMeterConfigActual(false);
     }
